Fixed using cidr notation
expand_cidr Method:

This method takes a CIDR notation (like 10.185.80.0/24) and expands it into individual IP addresses using the ipaddress module.
It handles errors if an invalid CIDR is provided.
Modified sniff_hosts:

When CIDR notation is detected in the hosts string, it expands the range using the expand_cidr method. This generates a list of individual IP addresses that can be scanned.